[opt] Optimize MyGetProcAddress

+++ fib_wrapper/fib_wrapper.cpp
@@ -2,7 +2,7 @@
 #include <windows.h>
 
 // Fib function signature
-typedef unsigned long (__stdcall* t_func)(int n);
+typedef unsigned long long (__stdcall* t_func)(int n);
 
 int main()
 {
